laravel - Done php file parsing, regex may not be perfect
